requirements.txt
# Framework web
Django==4.2.7
djangorestframework==3.14.0

# Base de datos
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Autenticación
django-allauth==0.57.0

# Seguridad
django-debug-toolbar==4.2.0
django-cors-headers==4.3.0

# Testing
pytest==7.4.3
pytest-django==4.6.0
pytest-cov==4.1.0

# Utilidades
python-dotenv==1.0.0
django-filter==23.3
Pillow==10.1.0
setuptools==68.2.2

# Documentación
drf-yasg==1.21.7

# Producción
gunicorn==21.2.0
whitenoise==6.5.0
reportlab==4.0.8 